implement n-parameter deobfuscation
Extract video and audio stream URLs from a YouTube link passed via
argv[1], and then print the results to stdout. This output can be
combined with an external video player like mpv:

    uri="$(xclip -o)"
    x="/tmp/streams.txt"
    youtube-unthrottle "$uri" > "$x"

    audio="$(head -1 $x)"
    video="$(tail -1 $x)"
    $mpv --title="$uri" --audio-file="$audio" "$video"

The main challenge here is that YouTube stream URLs contain a special
"n" GET parameter that must be deobfuscated using JavaScript fragments
supplied elsewhere in the YouTube payload. This is why solving this
puzzle requires the use of an embedded JavaScript engine (in this case,
Duktape).