split out landlock.c from sandbox.c
Idiomatic Landlock API usage seems to involve some preprocessor
spaghetti, probably to be forwards compatible with soon-to-be (but not
yet existent) libc wrappers for Landlock syscalls.

This makes it worthwhile to split landlock.c into its own file, in order
to make the "policy" aspects of sandbox.c more prominent. It's a bit of
an unnecessary layer of indirection, but I think it improves readability
and discoverability a bit, making the indirection worthwhile.