fix -fanalyzer warning re: uninitialized value
src/youtube.c:409:33: warning: use of uninitialized value [...] ciphertexts [CWE-457] [-Wanalyzer-use-of-uninitialized-value]

I'm hitting this compiler warning on an in-progress feature branch, but
the underlying logic error seems to exist on main branch as well.

I'm concerned that the intermediate `const size_t` is confusing the
compiler and (perhaps) contributing to this warning not reproducing on
main. Simplify by getting rid of the intermediate variable.