checkpoint seccomp filter after_inet()
seccomp filter can be pretty tight after curl is done, when the only
work remaining is to call into duktape and print the results.

A few approaches to consider:

Apply these restrictions as-is around duktape JS execution. This would
miss sandboxing the duktape JSON parsing functions though.

Separate the work into a group of coordinating processes, with all of
the duktape API calls happening in a process that restricted to read,
write, mmap, munmap, etc. That would allow the JSON parsing functions to
be sandboxed, even though they have to be called earlier, while the
program in general still needs to be able to call curl APIs.

Stay within a single process, but move JSON parsing later, after base.js
parsing. That is, first grab the base.js URL from the HTML blob and
trigger the base.js download via curl. Only after this is done and after
we've dropped privileges (like above), would we then parse the JSON
embedded in the HTML via duktape, which could happen in a lower
privilege context, like this commit does for duktape JS execution.