fix uninitialized value -> garbage into syscall()
I thought SECCOMP_GET_ACTION_AVAIL was setting <action>, but <action> is
actually read by the syscall. Failing to initialize the pointed-to value
was causing non-deterministic behavior, depending on what the compiler
decided to do with the variable on the stack.

Under clang, the sandbox-seccomp unit test was failing. Under gcc, it
was passing. This was all luck though.

Fix this issue by initializing the value explicitly.