add proof of origin token (poToken) support
Support poToken/visitorData challenge, based on:

    https://github.com/yt-dlp/yt-dlp/wiki/Extractors#po-token-guide
    https://github.com/yt-dlp/yt-dlp/pull/10648
    https://github.com/iv-org/invidious/pull/4789

By following the steps in the yt-dlp wiki page (linked above), I was
able to get a working (poToken, visitorData) pair in Chromium. For some
reason, I could not get a working pair in Firefox Nightly; maybe my
profile in the latter is fingerprinted and soft-blocked somehow?