run vale on prose of comments and docs
Use Vale to check the spelling and grammar of English prose in the
project repo, including block comments in C code and documentation in
Markdown README files.

References:

- https://lwn.net/Articles/964075/
- https://news.ycombinator.com/item?id=42407250
- https://lobste.rs/s/rupea8/3_shell_scripts_improve_your_writing_my_ph

Restricting comments in code and internal-only docs with Vale would
probably never make sense in a real software project, especially one
with a large team of varying language backgrounds.

That said, a pedantic approach does no harm for a personal project like
this; at worst, I only shoot myself in the foot.