extend build hack for OpenBSD ports to macOS brew
This work started as a set of experiments with library detection via
pkg-config. The general conclusion I took from this experimentation is
that it's best to stick with the current approach of assuming that the
system package manager has satisfied the project's dependencies and that
the developer's environment makes the relevant headers and libraries
available by default. Making this project's build system work well for
scenarios like cross-compiling, non-system-level dependencies, etc.
would be overengineering at this point.

That being said, I did realize during the work above that the existing
workarounds for /usr/local/{include,lib} on OpenBSD (aka dependencies
installed from ports) could be reasonably extended to make the configure
step on macOS simpler, avoiding the need to specify:

    CFLAGS="-isystem $(brew --prefix)/include"
    LDFLAGS="-L$(brew --prefix)/lib"

... when invoking cmake.

This commit implements the above idea, plus maintenance and readability
improvements inspired by the aforementioned experimentation on OpenBSD:

- change how CMakeLists.txt expresses dependencies within the project
    - use add_library() + target_link_libraries() idiom to grant
      test-driver access to internal implementation headers, rather than
      using target_include_directories(... src) directly
- mark third-party includes as SYSTEM in CMakeLists.txt, allowing us to
  drop linter/coverage regex exclusions for these header files
- wrap complex(-ish) cmake logic into modules
    - adjusting global include/library paths for Homebrew, OpenBSD ports
    - managing dependencies with get_cpm.cmake and CPMAddPackage()
    - generating C protobuf APIs from *.proto files
- standardize on PROJECT_SOURCE_DIR; drop CMAKE_SOURCE_DIR
- drop explicit CFLAGS/LDFLAGS/LD_LIBRARY_PATH workarounds on GitHub CI,
  as they are now unnecessary, just like in a devenv

This commit also includes a fix in ./scripts/qemu-build.exp for an
unrelated issue I happened to encounter on GitLab CI during this work.
By chance, I hit a timing issue where the version of glibc in the
nightly archlinux container image (host) exceeded the version of glibc
in the fortnightly archlinux VM image (guest), leading to process
crashes in the GitLab code-coverage job. Updating the latter from within
the VM, before running any tests, addresses the issue.